backend/requirements.txt
# ModEcon Matrix System Backend Requirements

# Web Framework
Flask>=2.0.1
Flask-RESTful>=0.3.9
Flask-Cors>=3.0.10

# Database
psycopg2-binary==2.9.10  # PostgreSQL adapter
clickhouse-driver>=0.2.0  # ClickHouse driver

# Utilities
python-dotenv>=0.19.1
numpy>=1.21.2
pandas>=1.3.3
scipy>=1.7.1
matplotlib>=3.4.3

# Testing
pytest>=6.2.5
pytest-cov>=2.12.1

# Documentation
Sphinx>=4.2.0
